fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! refactor(copyurl): remove redundant code
